// generated/src/app/api-models/smschedule-s.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SmscheduleS {
  namedobjectIdSequenceid: number;
  namedobjectIdObjecttype: number;
  scheduleHasexpiration: number;
  scheduleExpDate: Date;
  scheduleExpTodSpm: number;
  scheduleExpTodZone: number;
  repeatedIsimmediate: number;
  repeatedStarttimeDate: Date;
  repeatedStarttimeTodSpm: number;
  repeatedStarttimeTodZone: number;
  repeatedSecondsinperiod: number;
}